core: fixed remove operation in HdishDao

+++ restaurant-core/src/main/java/net/freedomserg/restaurant/core/model/hibernate/HdishDao.java
@@ -44,8 +44,10 @@ public class HdishDao implements DishDao {
         dish.setStatus(Status.DELETED);
         sessionFactory.getCurrentSession().update(dish);
         List<DishUnit> units = dish.getUnits();
-        for (DishUnit unit : units) {
-            dishUnitDao.remove(unit);
+        if (units != null) {
+            for (DishUnit unit : units) {
+                dishUnitDao.remove(unit);
+            }
         }
     }
